+++ packages/rules-unit-testing/test/database.test.ts
@@ -279,7 +279,7 @@ describe('Testing Module Tests', function () {
     // TODO: This test cannot be enabled without adding credentials to the test environment
     //       due to an underlying issue with firebase-admin storage. For now we will run it
     //       locally but not in CI.
-    if (process.env.CI !== "true") {
+    if (process.env.CI !== 'true') {
       await firebase.assertSucceeds(
         app.storage().bucket().file('/foo/bar.txt').save('Hello, World!')
       );